Making more progress on emulating PIF
As far as I know 95% of the CRC seems correct. Comparing the
memory to what someone else shows gives 15 out of 16 correct
memory results. I'm not sure why one is different, but I'll
move on and hope that a CPU test or something else later
gives a better indication of where to look.

As of now, it seems like CRC completes, but seems to enter
and infinite loop and the code never jumps to $A4000040 as
it's supposed to.