Spent the day tracking down a bug
This bug was preventing progress because it was corrupting memory and
register values.  It was a problem with the SxL/SxR and LxL/LxR
instructions. I had assumed that shifting a u32 by 32 bits (in either
direction) would result in the value 0, which in Rust appears to
actually be a shift by 0 (leaving the value alone).

The peripheral interface can print out debug info at the address used in
n64-systemtest.

I've also added a few more CPU instructions.