TLB exceptions and more
Finished a few more regimm instructions, too.

Uncommented out more of the TLB tests, which fail but so far the
systemtests are doing well.  I have 10 test cases that are commented out
because they prevent the tests from completing, and another 11 are
currently failing. So 21 or so failing tests out of 3000+ is 98-99%
passing tests.